add Service for Donation and view - number of all donated bags

+++ src/main/java/pl/lg/charity/services/DonationService.java
@@ -2,5 +2,5 @@ package pl.lg.charity.services;
 
 public interface DonationService {
 
-    Long findSumOfAllDonations();
+    Integer findSumOfAllDonations();
 }
